Move process method to trait
