+++ src/Commands/Launch.php
@@ -4,10 +4,12 @@ namespace A17\Blast\Commands;
 
 use Illuminate\Console\Command;
 use Illuminate\Filesystem\Filesystem;
-use Symfony\Component\Process\Process;
+use A17\Blast\Traits\Helpers;
 
 class Launch extends Command
 {
+    use Helpers;
+
     /**
      * The name and signature of the console command.
      *
@@ -128,29 +130,5 @@ class Launch extends Command
             'PROJECTPATH' => base_path(),
             'COMPONENTPATH' => base_path('resources/views/stories'),
         ]);
-    }
-
-    /**
-     * @return void
-     */
-    private function runProcessInBlast(
-        array $command,
-        $disableTimeout = false,
-        $envVars = null,
-    ) {
-        $process = new Process(
-            $command,
-            base_path($this->vendorPath),
-            $envVars,
-        );
-        $process->setTty(Process::isTtySupported());
-
-        if ($disableTimeout) {
-            $process->setTimeout(null);
-        } else {
-            $process->setTimeout(config('blast.build_timeout', 300));
-        }
-
-        $process->mustRun();
     }
 }
